refactor: variables management
