+++ apps/web/src/components/templates/email-editor/EmailMessagesCards.tsx
@@ -4,114 +4,17 @@ import { AuthContext } from '../../../store/authContext';
 import { When } from '../../utils/When';
 import { Preview } from '../../../pages/templates/editor/Preview';
 import { EditorPreviewSwitch } from '../EditorPreviewSwitch';
-import { Grid, Group, JsonInput, SegmentedControl, useMantineTheme } from '@mantine/core';
+import { Grid, Group, Modal, SegmentedControl, Title, useMantineTheme } from '@mantine/core';
 import { TestSendEmail } from './TestSendEmail';
-import { colors } from '../../../design-system';
+import { colors, shadows } from '../../../design-system';
 import { MobileIcon } from '../../../pages/templates/editor/PreviewSegment/MobileIcon';
 import { WebIcon } from '../../../pages/templates/editor/PreviewSegment/WebIcon';
-import { Collapse, UnstyledButton } from '@mantine/core';
-import { useFormContext, useWatch } from 'react-hook-form';
-import { SystemVariablesWithTypes } from '@novu/shared';
-import { ChevronUp } from '../../../design-system/icons';
-import { ChevronDown } from '../../../design-system/icons/arrows/ChevronDown';
+import { UnstyledButton } from '@mantine/core';
 import { ChevronRight } from '../../../design-system/icons/arrows/ChevronRight';
 import { ChevronLeft } from '../../../design-system/icons/arrows/ChevronLeft';
-import { inputStyles } from '../../../design-system/config/inputs.styles';
-import { useProcessVariables } from '../../../hooks/use-process-variables';
-
-const VarLabel = ({ label, children }) => {
-  const [open, setOpen] = useState(true);
-
-  return (
-    <>
-      <UnstyledButton
-        onClick={() => {
-          setOpen(!open);
-        }}
-        type="button"
-        sx={{
-          width: '100%',
-          marginBottom: 15,
-        }}
-      >
-        <div
-          style={{
-            color: colors.white,
-            fontSize: 14,
-            fontWeight: 'bold',
-          }}
-        >
-          {label}:
-          <span
-            style={{
-              float: 'right',
-            }}
-          >
-            {open ? <ChevronUp /> : <ChevronDown />}
-          </span>
-        </div>
-      </UnstyledButton>
-      <Collapse in={open}>{children}</Collapse>
-    </>
-  );
-};
-
-const VarItem = ({ name, type, children = null }: { name: string; type: string; children?: any }) => {
-  return (
-    <div
-      style={{
-        marginBottom: 10,
-        padding: 10,
-        borderRadius: 7,
-        background: colors.B20,
-        color: colors.B60,
-      }}
-    >
-      "{name}":"{type}"{children}
-    </div>
-  );
-};
-
-const VarItemsDropdown = ({ name, type }) => {
-  const [open, setOpen] = useState(false);
-
-  return (
-    <>
-      <UnstyledButton
-        onClick={() => {
-          setOpen(!open);
-        }}
-        type="button"
-        sx={{
-          width: '100%',
-        }}
-      >
-        <VarItem name={name} type="object">
-          <span
-            style={{
-              float: 'right',
-            }}
-          >
-            {open ? <ChevronUp /> : <ChevronDown />}
-          </span>
-        </VarItem>
-      </UnstyledButton>
-      <Collapse in={open}>
-        <div
-          style={{
-            borderBottom: `1px solid #292933`,
-            marginBottom: 10,
-            paddingLeft: 12,
-          }}
-        >
-          {Object.keys(type).map((key) => {
-            return <VarItem name={key} type={type[key]} />;
-          })}
-        </div>
-      </Collapse>
-    </>
-  );
-};
+import { useHotkeys } from '@mantine/hooks';
+import { VariableManager } from '../VariableManager';
+import { VariablesManagement } from './variables-management/VariablesManagement';
 
 export enum ViewEnum {
   EDIT = 'Edit',
@@ -132,14 +35,37 @@ export function EmailMessagesCards({
   const [view, setView] = useState<ViewEnum>(ViewEnum.EDIT);
   const [preview, setPreview] = useState<'mobile' | 'web'>('web');
   const theme = useMantineTheme();
-  const { control } = useFormContext();
   const [showVarManagement, setShowVarManagement] = useState(true);
-  const variableArray = useWatch({
-    name: `steps.${index}.template.variables`,
-    control,
-  });
+  const [modalOpen, setModalOpen] = useState(false);
 
-  const processedVars = useProcessVariables(variableArray);
+  useHotkeys([
+    [
+      '1',
+      () => {
+        setView(ViewEnum.EDIT);
+      },
+    ],
+    [
+      '2',
+      () => {
+        setView(ViewEnum.PREVIEW);
+        setPreview('web');
+      },
+    ],
+    [
+      '3',
+      () => {
+        setView(ViewEnum.TEST);
+      },
+    ],
+    [
+      '4',
+      () => {
+        setView(ViewEnum.PREVIEW);
+        setPreview('mobile');
+      },
+    ],
+  ]);
 
   return (
     <>
@@ -222,58 +148,42 @@ export function EmailMessagesCards({
               maxWidth: '325px',
             }}
           >
-            <div
-              style={{
-                width: '100%',
-                height: '100%',
-                background: colors.B17,
-                borderRadius: 7,
-                padding: 15,
+            <VariablesManagement
+              index={index}
+              openVariablesModal={() => {
+                setModalOpen(true);
               }}
-            >
-              <VarLabel label="System Vars">
-                {Object.keys(SystemVariablesWithTypes).map((name) => {
-                  const type = SystemVariablesWithTypes[name];
-
-                  if (typeof type === 'object') {
-                    return <VarItemsDropdown name={name} type={type} />;
-                  }
-
-                  return <VarItem name={name} type={type} />;
-                })}
-              </VarLabel>
-              <div
-                style={{
-                  marginTop: '20px',
-                }}
-              >
-                <VarLabel label="Step Vars">
-                  {variableArray.map((item) => {
-                    return <VarItem name={item.name} type={item.type.toLowerCase()} />;
-                  })}
-                </VarLabel>
-              </div>
-              <div
-                style={{
-                  marginTop: '20px',
-                }}
-              >
-                <VarLabel label="JSON">
-                  <JsonInput
-                    data-test-id="test-email-json-param"
-                    formatOnBlur
-                    autosize
-                    styles={inputStyles}
-                    value={processedVars}
-                    disabled
-                    minRows={6}
-                  />
-                </VarLabel>
-              </div>
-            </div>
+            />
           </Grid.Col>
         </Grid>
       </When>
+      <Modal
+        opened={modalOpen}
+        overlayColor={theme.colorScheme === 'dark' ? colors.BGDark : colors.BGLight}
+        overlayOpacity={0.7}
+        styles={{
+          modal: {
+            backgroundColor: theme.colorScheme === 'dark' ? colors.B15 : colors.white,
+            width: '90%',
+          },
+          body: {
+            paddingTop: '5px',
+            paddingInline: '8px',
+          },
+        }}
+        title={<Title>Variables</Title>}
+        sx={{ backdropFilter: 'blur(10px)' }}
+        shadow={theme.colorScheme === 'dark' ? shadows.dark : shadows.medium}
+        radius="md"
+        size="lg"
+        onClose={() => {
+          setModalOpen(false);
+        }}
+        centered
+        overflow="inside"
+      >
+        <VariableManager hideLabel={true} index={index} contents={['content', 'htmlContent', 'subject']} />
+      </Modal>
     </>
   );
 }
